devolutions-jet: start moving main to service structure
